examples/canary/requirements.txt
--extra-index-url https://pypi.nvidia.com
torch>=2.5.0
pytorch-lightning>=2.4.0
soundfile
safetensors
kaldialign
tensorrt_llm==0.17.0
datasets
lhotse
nemo==2.1.0
librosa